+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Npc skin cannot be applied
 - Not teleported to first camera in cutscene editor if no animations
 - Crash if going in cutscene editor with npcs without session set
+- Camera angle not taking custom global dialog data